# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News aggregator API documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/forgot-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "user reset password",
                "operationId": "e9dea57b981b47ff86a3786308ed9036",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@doe.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": [],
                                            "message": "A reset link has been sent to your email."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many attempts",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Too Many Attempts."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reset-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "user reset password",
                "operationId": "543c68a861604eb76de1e4934ec9884f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "token": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "email"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@doe.com",
                                    "token": "ki8191kaa",
                                    "password": "Pww123##pp"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": [],
                                            "message": "Your password has been updated successfully."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result-1": {
                                        "summary": "invalid token",
                                        "value": {
                                            "success": false,
                                            "data": null,
                                            "errors": {
                                                "token": [
                                                    "Reset link is invalid or has expired."
                                                ]
                                            }
                                        }
                                    },
                                    "result-2": {
                                        "summary": "invalid email",
                                        "value": {
                                            "success": false,
                                            "data": null,
                                            "errors": {
                                                "email": [
                                                    "Invalid email."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}